requirements.txt
absl-py==1.4.0
affine==2.4.0
altair==5.1.1
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
decorator==5.1.1
earthengine-api==0.1.369
exceptiongroup==1.1.3
executing==1.2.0
ffmpeg-python==0.2.0
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.42.1
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
h5py==3.9.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipyleaflet==0.17.4
ipython==8.15.0
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyterlab-widgets==3.0.9
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
pyproj==3.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
rasterio==1.3.8
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rioxarray==0.15.0
rpds-py==0.10.3
rsa==4.9
shapely==2.0.1
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
stack-data==0.6.2
streamlit==1.26.0
streamlit-folium==0.13.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.10.0
traittypes==0.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
xarray==2023.8.0
xyzservices==2023.7.0
zipp==3.16.2
